+++ app/src/main.rs
@@ -3,15 +3,16 @@
 
 use blind_signing::is_blind_signing_enabled;
 use blind_signing::update_blind_signing;
+use debug::print::println_array;
 use error_code::ErrorCode;
-use ledger_device_sdk::ecc::SeedDerive;
 use ledger_device_sdk::ui::layout;
 use ledger_device_sdk::ui::layout::Draw;
 use ledger_device_sdk::ui::layout::StringPlace;
 use ledger_secure_sdk_sys::buttons::ButtonEvent;
+use public_key::derive_pub_key;
 use sign_tx_context::SignTxContext;
 use tx_reviewer::TxReviewer;
-use utils::{self, deserialize_path, djb_hash, xor_bytes};
+use utils::{self, deserialize_path};
 mod blake2b_hasher;
 mod blind_signing;
 mod debug;
@@ -19,17 +20,15 @@ mod error_code;
 mod nvm_buffer;
 mod sign_tx_context;
 mod tx_reviewer;
+mod public_key;
 
-use debug::print::{println, println_array, println_slice};
-use ledger_device_sdk::ecc::{ECPublicKey, Secp256k1};
+use debug::print::{println, println_slice};
 use ledger_device_sdk::io;
 use ledger_device_sdk::ui::bagls;
 use ledger_device_sdk::ui::gadgets;
 use ledger_device_sdk::ui::screen_util;
 
 ledger_device_sdk::set_panic!(ledger_device_sdk::exiting_panic);
-
-const TOTAL_NUMBER_OF_GROUPS: u8 = 4;
 
 fn show_ui_common(draw: fn() -> ()) {
     gadgets::clear_screen();
@@ -179,18 +178,6 @@ impl TryFrom<io::ApduHeader> for Ins {
 
 use ledger_device_sdk::io::Reply;
 
-use crate::blake2b_hasher::Blake2bHasher;
-
-fn check_group(p1: u8, p2: u8) -> Result<(), Reply> {
-    if p1 == 0 && p2 == 0 {
-        return Ok(());
-    }
-    if p2 >= p1 || p1 != TOTAL_NUMBER_OF_GROUPS {
-        return Err(ErrorCode::BadP1P2.into());
-    }
-    return Ok(());
-}
-
 fn handle_apdu(
     comm: &mut io::Comm,
     ins: Ins,
@@ -225,16 +212,7 @@ fn handle_apdu(
 
             let p1 = apdu_header.p1;
             let p2 = apdu_header.p2;
-            check_group(p1, p2)?;
-
-            let (pk, hd_index) = if p1 == 0 {
-                (derive_pub_key(&path)?, path[path.len() - 1])
-            } else {
-                let group_num = p1;
-                let target_group = p2 % p1;
-                assert!(target_group < group_num);
-                derive_pub_key_for_group(&mut path, group_num, target_group)?
-            };
+            let (pk, hd_index) = derive_pub_key(&mut path, p1, p2)?;
 
             println_slice::<130>(pk.as_ref());
             comm.append(pk.as_ref());
@@ -265,56 +243,4 @@ fn handle_apdu(
         }
     }
     Ok(false)
-}
-
-fn derive_pub_key(path: &[u32]) -> Result<ECPublicKey<65, 'W'>, Reply> {
-    let pk = Secp256k1::derive_from_path(path)
-        .public_key()
-        .map_err(|x| Reply(0x6eu16 | (x as u16 & 0xff)))?;
-    Ok(pk)
-}
-
-fn derive_pub_key_for_group(
-    path: &mut [u32],
-    group_num: u8,
-    target_group: u8,
-) -> Result<(ECPublicKey<65, 'W'>, u32), Reply> {
-    loop {
-        println("path");
-        println_slice::<8>(&path.last().unwrap().to_be_bytes());
-        let pk = derive_pub_key(path)?;
-        if get_pub_key_group(pk.as_ref(), group_num) == target_group {
-            return Ok((pk, path[path.len() - 1]));
-        }
-        path[path.len() - 1] += 1;
-    }
-}
-
-pub fn get_pub_key_group(pub_key: &[u8], group_num: u8) -> u8 {
-    assert!(pub_key.len() == 65);
-    println("pub_key 65");
-    println_slice::<130>(pub_key);
-    let mut compressed = [0_u8; 33];
-    compressed[1..33].copy_from_slice(&pub_key[1..33]);
-    if pub_key.last().unwrap() % 2 == 0 {
-        compressed[0] = 0x02
-    } else {
-        compressed[0] = 0x03
-    }
-    println("compressed");
-    println_slice::<66>(&compressed);
-
-    let pub_key_hash = Blake2bHasher::hash(&compressed).unwrap();
-    println("blake2b done");
-    let script_hint = djb_hash(&pub_key_hash) | 1;
-    println("hint done");
-    let group_index = xor_bytes(script_hint);
-    println("pub key hash");
-    println_slice::<64>(&pub_key_hash);
-    println("script hint");
-    println_slice::<8>(&script_hint.to_be_bytes());
-    println("group index");
-    println_slice::<2>(&[group_index]);
-
-    group_index % group_num
 }
